requirements.txt
fastapi
uvicorn
celery
pika
pillow
python-multipart
